REG-7 - exclude IE11 from pledge form test
Xpath expressions needed to select labels appropriately
don't work in IE11 even though the form loads there.

As we already warn it's unsupported and will probably
phase out Donate tests by the end of this year, skipping
the browser and running this test only in Chrome
for now seems reasonable.