Delete failing feature Allocate champion funds to a charity successfully
This is failing for reasons that seem to be unrelated to any work we
have done recently, and I can't imediatly see how to fix it.

But we need to either fix it or remove it otherwise we'll be constantly
notified that the test suite is failing and we won't notice if any other
features start to fail after we push a breaking change.