Comment out "When I press on the button to set a password" step for now
This part was left in but doesn't test anything meaningful without the
subsequent lines commented while working on REG-33. And it seems to be failing
~80-90% of the time now for some reason.

To actually fix it we probably need a delay, an ID on the custom component,
a selector targeting the actual button inside it, upgrades to the libraries
involved in building or targeting the shadow DOM, or some combination of
these.