DON-965: Make tests clearer by removing code to close cookie notice
If we want these feature files to be real BDD style tests that we can
ask non-developers to read then we need to make sure they express the
core requriements as clearly as possible and have little to no
accidental complexity reflected in the actual feature files.

Since frontend is now specifically not showing cookie notices in
the regression environment we don't need to deal with them here.