MAT-384: Use hard-coded donation amounts in most tests
I think hard-coded amounts make the tests easier to write and clearer to
read and verify, and more relvantly here mean that when we come to
include a check of the fee calculation we won't need to repeat the logic
for calculating the fee in the test - we can calculate it once and
hard-code it as well.

Using £25k as the amount since if that works its likely anything smaller
will. If we wanted to test as thourghaly as possible we'd also have a
test for £1.